Ref: Created /search_filter_categories and moved files inside
