+++ app/src/components/market/sections/common_sections/search_filter_categories/categories/index.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react'
 import styled from 'styled-components'
 
-import { CATEGORIES } from '../../../../common/constants'
-import { Button } from '../../../button'
-import { ButtonType } from '../../../button/button_styling_types'
-import { Textfield } from '../../../common/'
+import { CATEGORIES } from '../../../../../../common/constants'
+import { Button } from '../../../../../button'
+import { ButtonType } from '../../../../../button/button_styling_types'
+import { Textfield } from '../../../../../common'
 
 interface Props {
   categories: string[]
