+++ app/src/components/market/sections/market_create/steps/items/ask_question_step.tsx
@@ -13,11 +13,11 @@ import { DateField, FormRow, FormStateButton } from '../../../../../common'
 import { BigNumberInputReturn } from '../../../../../common/form/big_number_input'
 import { CommonDisabledCSS } from '../../../../../common/form/common_styled'
 import { QuestionInput } from '../../../../../common/form/question_input'
-import { Categories } from '../../../../common/categories'
 import { MarginsButton } from '../../../../common/common_styled'
 import { Arbitrators } from '../../../common_sections/arbitrators/arbitrators'
 import { CreateCard } from '../../../common_sections/create_card'
 import { WarningMessage } from '../../../common_sections/message_text/warning_message'
+import { Categories } from '../../../common_sections/search_filter_categories/categories'
 import { Outcome, Outcomes } from '../outcomes'
 
 import { CreateScalarMarket } from './create_scalar_market'
